Added more Mechanic Constants, based type sizing off those constants, and used those types in more places

+++ src/Types/Enums/Gender.hpp
@@ -1,11 +1,10 @@
 #pragma once
 
-#include <cstddef>
 #include <cstdint>
 
 namespace pokesim::dex {
 // Pokemon gender name
 enum class Gender : std::uint8_t { NO_GENDER = 0, FEMALE, MALE };
 
-static constexpr std::size_t TOTAL_GENDER_COUNT = 3U;
+static constexpr std::uint8_t TOTAL_GENDER_COUNT = 3U;
 }  // namespace pokesim::dex
