+++ src/Types/Enums/Nature.hpp
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <cstddef>
 #include <cstdint>
 
 namespace pokesim::dex {
@@ -11,5 +10,5 @@ enum class Nature : std::uint8_t {
   // clang-format on
 };
 
-static constexpr std::size_t TOTAL_NATURE_COUNT = (std::size_t)Nature::NATURE_TOTAL - 1U;
+static constexpr std::uint8_t TOTAL_NATURE_COUNT = (std::uint8_t)Nature::NATURE_TOTAL - 1U;
 }  // namespace pokesim::dex
